fixes #51: change to pipeline insertion script
fixes #51: Move insert_new_version_pipeline to pipelines_version_insertion.py
