+++ azure_storage/azure_storage_api.py
@@ -276,40 +276,3 @@ async def get_pipeline_info(
 
     except (ValueError, GetBlobError, PipelineNotFoundError) as error:
         raise PipelineNotFoundError(f"This version {pipeline_version} was not found") from error
-
-
-def insert_new_version_pipeline(
-        pipelines_json: dict,
-        connection_string: str,
-        pipeline_container_name: str
-    ) -> bool:
-    """
-    Inserts a new version of a pipeline JSON into an Azure Blob Storage
-    container.
-
-    Args:
-        pipelines_json (dict): The JSON data of the pipeline. connection_string
-        (str): The connection string for the Azure Blob Storage account.
-        pipeline_container_name (str): The name of the container where the
-        pipeline JSON will be uploaded.
-
-    Returns:
-        bool: True if the pipeline JSON was successfully uploaded, False
-        otherwise.
-    """
-    try:
-        blob_service_client = BlobServiceClient.from_connection_string(
-            connection_string
-        )
-
-        container_client = blob_service_client.get_container_client(
-            pipeline_container_name
-        )
-
-        json_name = "{}/{}.json".format("pipelines", pipelines_json.get("version"))
-        container_client.upload_blob(
-            json_name, json.dumps(pipelines_json, indent=4), overwrite=True)
-        return "The pipeline was successfully uploaded to the blob storage"
-
-    except ValueError as error:
-        raise ConnectionStringError(error.args[0]) from error
